refactor(example): use barrel import for auth module (#1808)
