+++ projects/example-app/src/app/auth/actions/login-page.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { Credentials } from '@example-app/auth/models/user';
+import { Credentials } from '@example-app/auth/models';
 
 export const login = createAction(
   '[Login Page] Login',
