+++ projects/example-app/src/app/auth/auth-routing.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
-import { LoginPageComponent } from '@example-app/auth/containers/login-page.component';
+import { LoginPageComponent } from '@example-app/auth/containers';
 
 const routes: Routes = [{ path: 'login', component: LoginPageComponent }];
