+++ projects/example-app/src/app/auth/components/index.ts
@@ -1,0 +1,2 @@
+export * from './login-form.component';
+export * from './logout-confirmation-dialog.component';
