+++ projects/example-app/src/app/auth/components/login-form.component.spec.ts
@@ -1,6 +1,6 @@
 import { TestBed, ComponentFixture } from '@angular/core/testing';
 import { NO_ERRORS_SCHEMA } from '@angular/core';
-import { LoginFormComponent } from '@example-app/auth/components/login-form.component';
+import { LoginFormComponent } from '@example-app/auth/components';
 import { ReactiveFormsModule } from '@angular/forms';
 
 describe('Login Page', () => {
