+++ projects/example-app/src/app/auth/containers/login-page.component.spec.ts
@@ -3,8 +3,8 @@ import { MatInputModule, MatCardModule } from '@angular/material';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { Store } from '@ngrx/store';
-import { LoginPageComponent } from '@example-app/auth/containers/login-page.component';
-import { LoginFormComponent } from '@example-app/auth/components/login-form.component';
+import { LoginPageComponent } from '@example-app/auth/containers';
+import { LoginFormComponent } from '@example-app/auth/components';
 import * as fromAuth from '@example-app/auth/reducers';
 import { LoginPageActions } from '@example-app/auth/actions';
 import { provideMockStore, MockStore } from '@ngrx/store/testing';
