+++ projects/example-app/src/app/auth/effects/auth.effects.spec.ts
@@ -11,9 +11,9 @@ import {
   AuthApiActions,
 } from '@example-app/auth/actions';
 
-import { Credentials, User } from '@example-app/auth/models/user';
-import { AuthService } from '@example-app/auth/services/auth.service';
-import { AuthEffects } from '@example-app/auth/effects/auth.effects';
+import { Credentials, User } from '@example-app/auth/models';
+import { AuthService } from '@example-app/auth/services';
+import { AuthEffects } from '@example-app/auth/effects';
 
 describe('AuthEffects', () => {
   let effects: AuthEffects;
