+++ projects/example-app/src/app/auth/reducers/login-page.reducer.spec.ts
@@ -3,7 +3,7 @@ import * as fromLoginPage from '@example-app/auth/reducers/login-page.reducer';
 
 import { AuthApiActions, LoginPageActions } from '@example-app/auth/actions';
 
-import { Credentials, User } from '@example-app/auth/models/user';
+import { Credentials, User } from '@example-app/auth/models';
 
 describe('LoginPageReducer', () => {
   describe('undefined action', () => {
@@ -21,11 +21,6 @@ describe('LoginPageReducer', () => {
       const user = { username: 'test' } as Credentials;
       const createAction = LoginPageActions.login({ credentials: user });
 
-      const expectedResult = {
-        error: null,
-        pending: true,
-      };
-
       const result = reducer(fromLoginPage.initialState, createAction);
 
       expect(result).toMatchSnapshot();
@@ -37,11 +32,6 @@ describe('LoginPageReducer', () => {
       const user = { name: 'test' } as User;
       const createAction = AuthApiActions.loginSuccess({ user });
 
-      const expectedResult = {
-        error: null,
-        pending: false,
-      };
-
       const result = reducer(fromLoginPage.initialState, createAction);
 
       expect(result).toMatchSnapshot();
@@ -52,11 +42,6 @@ describe('LoginPageReducer', () => {
     it('should have an error and no pending state', () => {
       const error = 'login failed';
       const createAction = AuthApiActions.loginFailure({ error });
-
-      const expectedResult = {
-        error: error,
-        pending: false,
-      };
 
       const result = reducer(fromLoginPage.initialState, createAction);
