+++ projects/example-app/src/app/auth/services/index.ts
@@ -1,0 +1,2 @@
+export * from './auth.service';
+export * from './auth-guard.service';
